+++ scglobals/static/tx/tests/index.js
@@ -8,24 +8,22 @@ import Router from '../src/app/router';
 
 
 describe('Integration tests', function() {
-  debugger;
+  $('body').append('<div id="supplychain"></div>');
   const router = new Router();
   Backbone.history.start();
-  let $app = $('#js-app');
-
-
-  describe('Dashboard page sample testing', function() {
-    it('should render hello ignacio on first render', function() {
-      router.dashboard();
-      debugger;
-      expect($app.text()).to.equal('is this a thing? !');
+  let $app = $('#supplychain');
+  describe.only('Dashboard', function() {
+    it('should render the header', function() {
+      expect($app.text().toLowerCase()).to.contain('supply chain');
     });
+    // it('should render the header', function() {
+    //   expect($app.text().toLowerCase()).to.contain('supply chain');
+    // });
   });
 
   describe('About page sample testing', function() {
     it('should render YO im the about page', function() {
       router.navigate('about', {trigger: true, replace: true});
-
       expect($app.text()).to.equal('YO im the about page');
     });
   });